+++ CHANGELOG.md
@@ -72,6 +72,7 @@ For older changes see the [archived Singularity change log](https://github.com/a
   `<library>.so[.version]` files are listed by `ldconfig -p`.
 - Fix joining cgroup of instance started as root, with cgroups v1,
   non-default cgroupfs manager, and no device rules.
+- Ensure `DOCKER_HOST` is honored in non-build flows.
 
 ### Bug fixes
 
@@ -80,7 +81,6 @@ For older changes see the [archived Singularity change log](https://github.com/a
 - Avoid `unknown option` error when using a bare squashfs image with
   an unpatched `squashfuse_ll`.
 - Fix `GOCACHE` settings for golang build on PPA build environment.
-- Ensure `DOCKER_HOST` is honored in non-build flows.
 - Make the error message more helpful in another place where a remote is found
   to have no library client.
 - Allow symlinks to the compiled prefix for suid installations.  Fixes a