+++ CHANGELOG.md
@@ -16,7 +16,9 @@ For older changes see the [archived Singularity change log](https://github.com/a
 - Fix `GOCACHE` settings for golang build on PPA build environment.
 - Make the error message more helpful in another place where a remote is found
   to have no library client.
-- Avoid incorrect error when reqesting fakeroot network.
+- Allow symlinks to the compiled prefix for suid installations.  Fixes a
+  regression introduced in 1.1.4.
+- Avoid incorrect error when requesting fakeroot network.
 
 ## v1.1.6 - \[2023-02-14\]